Application domain fix
