+++ src/main/java/smartpot/com/api/Models/DAO/RCrop.java
@@ -1,8 +1,8 @@
-package smarpot.com.api.Models.DAO;
+package smartpot.com.api.Models.DAO;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-import smarpot.com.api.Models.Entity.Crop;
+import smartpot.com.api.Models.Entity.Crop;
 
 @Repository
 public interface RCrop extends MongoRepository<Crop, String> {
