+++ src/main/java/smartpot/com/api/Models/DAO/RSession.java
@@ -1,8 +1,8 @@
-package smarpot.com.api.Models.DAO;
+package smartpot.com.api.Models.DAO;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-import smarpot.com.api.Models.Entity.Session;
+import smartpot.com.api.Models.Entity.Session;
 
 @Repository
 public interface RSession extends MongoRepository<Session, String> {
