+++ src/main/java/smartpot/com/api/Models/DAO/RUser.java
@@ -1,6 +1,6 @@
-package smarpot.com.api.Models.DAO;
+package smartpot.com.api.Models.DAO;
 
-import smarpot.com.api.Models.Entity.User;
+import smartpot.com.api.Models.Entity.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
