+++ src/main/java/smartpot/com/api/Models/Entity/Crop.java
@@ -1,6 +1,6 @@
-package smarpot.com.api.Models.Entity;
+package smartpot.com.api.Models.Entity;
 
-import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -8,19 +8,21 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.util.Date;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Document(collection = "sesiones")
-public class Session {
+@Document(collection = "cultivos")
+public class Crop {
     @Id
     private String id;
 
-    @NotNull(message = "La fecha de registro no puede estar vacía")
-    private Date registration;
+    private String status;
+
+    @NotEmpty(message = "El tipo no puede estar vacío")
+    private String type;
 
     private String user;
+
+    private String cultivation;
 }
