+++ src/main/java/smartpot/com/api/Models/Entity/History.java
@@ -1,4 +1,4 @@
-package smarpot.com.api.Models.Entity;
+package smartpot.com.api.Models.Entity;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
