+++ src/main/java/smartpot/com/api/Models/Entity/Session.java
@@ -1,6 +1,6 @@
-package smarpot.com.api.Models.Entity;
+package smartpot.com.api.Models.Entity;
 
-import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -8,21 +8,19 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.util.Date;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Document(collection = "cultivos")
-public class Crop {
+@Document(collection = "sesiones")
+public class Session {
     @Id
     private String id;
 
-    private String status;
-
-    @NotEmpty(message = "El tipo no puede estar vacío")
-    private String type;
+    @NotNull(message = "La fecha de registro no puede estar vacía")
+    private Date registration;
 
     private String user;
-
-    private String cultivation;
 }
