+++ src/main/java/smartpot/com/api/SmartPotApiApplication.java
@@ -1,11 +1,11 @@
-package smarpot.com.api;
+package smartpot.com.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 
 @SpringBootApplication
-@EnableMongoRepositories(basePackages = "smarpot.com.api")
+@EnableMongoRepositories(basePackages = "smartpot.com.api")
 public class SmartPotApiApplication {
 
     public static void main(String[] args) {
