+++ src/test/java/smartpot/com/api/SmartPotApiApplicationTests.java
@@ -1,4 +1,4 @@
-package smarpot.com.api;
+package smartpot.com.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
